Split functions as modules & pycache ignore
